% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boston.R
\docType{data}
\name{boston}
\alias{boston}
\title{Boston}
\format{
data.frame, 506 observations of 14 variables:
\describe{
  \item{crim}{per capita crime rate by town.}
  \item{zn}{proportion of residential land zoned for lots over 25,000 sq.ft.}
  \item{indus}{proportion of non-retail business acres per town.}
  \item{chas}{Charles River dummy variable (= 1 if tract bounds river; 0
        otherwise).}
  \item{nox}{nitrogen oxides concentration (parts per 10 million).}
  \item{rm}{average number of rooms per dwelling.}
  \item{age}{proportion of owner-occupied units built prior to 1940.}
  \item{dis}{weighted mean of distances to five Boston employment centers.}
  \item{rad}{index of accessibility to radial highways.}
  \item{tax}{full-value property-tax rate per $10,000.}
  \item{ptratio}{pupil-teacher ratio by town.}
  \item{black}{1000(Bk - 0.63)^2 where Bk is the proportion of blacks by
        town.}
  \item{lstat}{lower status of the population (percent).}
  \item{medv}{median value of owner-occupied homes in $1000s.}
}
}
\usage{
data("boston")
}
\description{
Boston housing data set
}
\keyword{datasets}
