% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_pa_titanic.R
\docType{data}
\name{exam_pa_titanic}
\alias{exam_pa_titanic}
\title{Exam PA Titanic}
\format{
data.frame, 906 observations of 11 variables:
\describe{
  \item{passengerid}{Passenger id}
  \item{survived}{Survived Y/N}
  \item{pclass}{Ticket class}
  \item{name}{Name}
  \item{sex}{male, female}
  \item{age}{Age}
  \item{sibsp}{# of siblings}
  \item{parch}{# of parents or children aboard the Titanic.}
  \item{ticket}{Number fare}
  \item{fare}{Cost of ticket.}
  \item{embarked}{Port of Embarkation. C = Cherbourg, Q = Queenstown,
        S = Southampton.}
}
}
\usage{
exam_pa_titanic
}
\description{
Titanic passengers as used in PredictiveInsightsAI.com’s practice exam
}
\keyword{datasets}
