% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_insurance.R
\docType{data}
\name{health_insurance}
\alias{health_insurance}
\title{Health insurance}
\format{
data.frame, 1338 observations of 7 variables:
\describe{
  \item{age}{Age of policy holder.}
  \item{sex}{M or F.}
  \item{bmi}{Body Mass Index: weight divided by height.}
  \item{children}{Number of children.}
  \item{smoker}{Smoker status. Yes or No.}
  \item{region}{Geographic region.}
  \item{charges}{Annual medical claims for this policy.}
}
}
\usage{
data("health_insurance")
}
\description{
Health insurance claims as used in ExamPA.net’s Practice Exam.
The data set consists of prior year’s health insurance claims, along with
patient demographic information, from Freedom Health.
}
\keyword{datasets}
