% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_num_labs.R
\docType{data}
\name{patient_num_labs}
\alias{patient_num_labs}
\title{Patient Number of Labs}
\format{
data.frame,  10000 observations of 14 variables:
\describe{
  \item{age}{Age of prospective customer. Integer from 17 to 98.}
  \item{job}{Occupation category. Factor with 11 levels.}
  \item{marital}{Marital status. Factor with levels divorced, married, and
        single}
  \item{housing}{Indicates whether the prospect has a housing loan. Factor
        with levels no, yes.}
  \item{loan}{Indicates whether the prospect has a consumer loan. Factor with
        levels no, yes.}
  \item{phone}{The type of phone the prospect uses. Factor with levels
        cellular, landline.}
  \item{month}{The month of the marketing call. Factor with 12 levels.}
  \item{weekday}{The day of the week of the marketing call. Factor with five
        levels.}
  \item{CPI}{Consumer price index at the time of the call. Numeric from
        92.20 to 94.77.}
  \item{CCI}{Consumer confidence index at the time of the call. Numeric
        from -50.8 to -26.9.}
  \item{irate}{Short term interest rate at the tie of the call. Numeric
        from 0.634 to 5.045.}
  \item{employment}{Number of employees of ABC Insurance at the time of the
        call. Numeric from 4964 to 5228.}
  \item{purchase}{Indicator of purchase. Integer (1 for purchase, or 0 for
        no purchase.)}
  \item{edu_years}{Years of education. Integer from 1 to 16.}
}
}
\usage{
data("patient_num_labs")
}
\description{
Data used on June 19, 2020 Exam PA
}
\keyword{datasets}
