% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedestrian_activity.R
\docType{data}
\name{pedestrian_activity}
\alias{pedestrian_activity}
\title{Pedestrian activity}
\format{
data.frame, 11373 observations of 7 variables:
\describe{
  \item{pedestrians}{The count of pedestrians during one hour starting at the indicated time.}
  \item{weather}{Hourly weather condition, eleven categories.}
  \item{temperature}{Hourly temperature in degrees Fahrenheit. }
  \item{precipitation}{Hourly precipitation in inches.}
  \item{hour}{Time at beginning of the measuring hour.}
  \item{weekday}{Day of the week.}
  \item{temp_forecast}{Predicted daily average temperature in degrees Fahrenheit.}
}
}
\usage{
pedestrian_activity
}
\description{
pedestrian activity dataset
}
\keyword{datasets}
