% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_insurance.R
\docType{data}
\name{travel_insurance}
\alias{travel_insurance}
\title{Travel insurance data}
\format{
data.frame, 10000 observations of 7 variables:
\describe{
  \item{Distance}{Distance traveled in trip, in km}
  \item{Duration}{Number of nights spent on trip}
  \item{Reason}{Main reason for the trip. Vacation includes holiday, leisure, or recration. Visit includes visiting friends or relatives.}
  \item{Age}{Age of adult survey respondent}
  \item{Others}{Number of other persons that accompanied the respondent on the trip}
  \item{Mode}{Main mode of transportation, car or plane}
  \item{Cost}{Total spending on trip, in Canadian $}
}
}
\usage{
travel_insurance
}
\description{
The travel insurance dataset.
}
\keyword{datasets}
