% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_spending.R
\docType{data}
\name{travel_spending}
\alias{travel_spending}
\title{Travel spending data}
\format{
data.frame, 4884 observations of 11 variables:
\describe{
  \item{Q}{Calender quarter of trip}
  \item{ProvO}{Trip province of origin}
  \item{Distance}{Distance traveled in trip, in km}
  \item{Duration}{Number of nights spent on trip}
  \item{Reason}{Main reason for the trip. Vacation includes holiday, leisure, or recration. Visit includes visiting friends or relatives.}
  \item{Age}{Age of adult survey respondent in six age bins}
  \item{Gender}{Gender of adult survey respondent}
  \item{HHI}{Household income, in Canadian $}
  \item{Others}{Number of other persons that accompanied the respondent on the trip}
  \item{Mode}{Main mode of transportation, car or plane}
  \item{Cost}{Total spending on trip, in Canadian $}
}
}
\usage{
travel_spending
}
\description{
The travel spending dataset.
}
\keyword{datasets}
